Major simplification of lamp effect module.  Changes include:
- Added leff_running_p() to query if a lamp effect is started (it may not be visible though).
Fast because queued effects use a bitset to track them, instead of a
dynamic queue.

- Number of queued effects matches total number of leffs defined by machine.
Before it could only be 8.

- Starting an effect already started, or vice-versa, is a no-op.  This
would cause unpredictable behavior before.

- The concept of _the_ running effect is gone.  There is no such thing.
One may only be visible, but what is visible is irrelevant.

- Checking for high priority when starting a new effect is much more
optimal than before, since we keep the highest priority started in
a variable.  It does not need to be recomputed.

- The default leff is gone.  If no lamp effect is started, there is no
task running at all now.  Before a dummy task was running.