Many changes to get sound working on real hardware:
- Sound init divided into two parts: one that runs early
while interrupts are still disabled, and another that
runs later when task scheduling is working.
Previously, what was running at task level could not
actually be scheduled properly.

- Sound board version string will now report if there
was any error.

- Sound board commands for DCS supported, although not
fully tested.  Does not appear to be working for
init.

- Fix race condition when entering test mode.  The
enter sound was not always getting played because
the sound cutoff hadn't finished yet.