Initial version of solenoid request manager.
The idea is to replace the repetitive checks in the IRQ on each
coil, with a single check that services multiple, nonoverlapping
coils.  At the same time, we will add support for zerocross
detection so that firing can be initiated near a voltage peak for
maximum "punch", and support two-phase (strong then duty) pulsing
to give a strong blast initially but tone it down thereafter.