Teach gencallset about CALLSET_ENTRYs that are used for more than one input.
Previously, the second argument was an event name, and only
two arguments could be given.  Now, _at least_ two arguments must
be given, allowing for a list of events.  For example,

CALLSET_ENTRY (mymodule, x, y, z)

would declare a function to be invoked whenever x, y, or z was
thrown.

The name of the generated function only uses the first event,
in order to keep symbol names short.