Change the way that alphanumeric games are defined.
Previously, if MACHINE_DMD was not defined, then alphanumeric was
implied.  However, there are more than two options; we are now
able to account better for Whitestar with remote DMD, or P-ROC
with a different form of DMD entirely.  So alphanumeric display
must now be configured explicitly.

The Alphanumeric: Yes line in the global config enables it,
and generates MACHINE_ALPHANUMERIC = 1 to be used throughout
the C code; Makefiles use CONFIG_ALPHA.

If neither DMD nor ALPHA is defined, the DMD APIs are now macros
that do nothing; this helps transition code without major breakage.