Simplified the way lamp effects are classified.
The L_NORMAL state (neither L_RUNNING nor L_SHARED set) was not
being used for anything.  The default lamp effect was changed to
be of type L_RUNNING.  genmachine now considered a leff to be
running by default.  The 'runner' property still works for
compatibility.

Later, the meaning of these flags should be changed; we don't
need two bit values for two mutually exclusive things.  We should
make this as lamp effect ID, with zero meaning 'running' or
global as today and different nonzero value for each shared leff.