Implement the "Chase Ball" adjustment.
When enabled, after the 5th unsuccessful ball search, the game will
end the current ball, flag all balls not accounted for as missing,
and continue with the next ball (or end the game).  This is to
prevent infinite ball search.

Default is now for chase ball to be OFF, however.  It can be
enabled via the adjustment.