Remove KLOCK_DEBUGGER.
This was defined to allow the inspector to pause games while debugging.
The KLOCK_DEFF flag already does that when a pausing deff is running,
which is what the inspector is.  So set D_PAUSE on it and use the same
lock flag.

Untested, as the inspector has not been in the builds for a while.