TZ : Redesign mechanical clock driver.
This change is necessary due to changes in the way that underlying
solenoid drivers are now being handled.  It also takes advantage of
the new template drivers that were not available when the clock driver
was first written.

At the lowest layer is the clock motor driver, which is based on the
'bivar' template (for bidirectional, variable-speed).  This driver
is for a 3-state device (off, forward, reverse) and the speed is
controllable.

Next is another realtime function which handles the clock switches.
This keeps tracking of the clock hand positions based on the switch
states.  It also allows the clock to be stopped instantly when the
current time reaches a fixed value (currently only used to get home
at 12:00)

Finally is a slow update function that does not run realtime.  It
handles timeouts, error checking, and state machine transitions.