Fix a subtle race condition in the ball tracker, where a switch closure in a device is not handled at all.
What happened was a switch closed while the device update
function was running, but after it had polled the device count
for the final time.  Switch closures are ignored while update
is already running.

The problem would not normally occur, because update and
switch processing should not conflict, unless there is
preemption in between them.  The culprit seen was
device_debug(), which had a small delay inserted to avoid
task lockup during the huge write to the debug port.
But it is not inconceivable for device-related callbacks to
do some sleeping as well.  So the problem needs to be solved
generally.

The solution is to repoll the switches one last time, just
before exiting the task.  Within this code block, *no*
task switch should happen.  If we see a change in the switch
count here, then we go back to the top of the update loop,
just as if the switch handler had invoked it.

Along with this core change, the debug functions were
simplified; we no longer print out the state of all devices,
only those that are being updated.  This makes it much
easier to understand what is happening.