Move platform-specific defines from game configs to platform configs.
Previously, each game had to include a platform config AND declare
whether it used DMD, PIC, etc., which would be the same for every
game on that platform.  So those definitions are now moved into the
file that was already being included.

Some shortcuts had been taken and the proper config was not being
included in all places; that has been fixed.