Device driver outputs are now smoothed over 2ms intervals.
Previously, when a driver did sol_enable() or sol_disable(),
the effect was immediate.  During the course of an interrupt,
many drivers would do this and the physical I/O register was
being written many times.

Now, the device driver only updates the in-memory state of
the solenoid bank output.  A separate realtime function
copies this to the hardware at a less frequent rate.  This
saves a few cycles, but it does not permit duty cycling
faster than every 2ms.