Optimize WPC fliptronic driver.
This patch does not change the overall algorithm used, which is still known
to have issues and is in need of improvement for robustness.  Specific
changes here are:

- Use __fastram__ on many variables needed.

- Use global variables for some temporaries rather than having GCC try
to place them in registers.  Ultimately this was causing a lot of stack
variables which were inefficient.

- Change the single flipper update function flipper_service() not to
use pointers anymore.  GCC6809 could not optimize well wehn written
this way.

- Add a check for the specific case when no flipper switches (EOS or
buttons) are active.  In this case no outputs should be turned on.
Previously we were checking each flipper in turn, but this is a
common case.

- Now, we also turn on the holding coil whenever the power coil is on.

- Use likely() and unlikely() around many conditions, to help the
compiler optimize branches.