Add support for dynamic breakpoints.
APIs are added to set/clear a single breakpoint address, which
must be inside a dynamic task (no interrupt handler or
periodic function).  All potential breakpoint locations must
be marked in the code with the bpt() macro, which will check
if the current address matches the desired break address.

On a breakpoint hit, task scheduling and most periodic
functions are suspended.  To continue, press the start button,
or if using the emulation debugger console, press the 'p'
key.  Interrupt-level functions cannot be stopped.  The
periodic routine for switch polling still runs, in order
to determine when start has been pressed to continue.