Initial checkin of the debug shell.
This feature requires that DEBUGGER and CONFIG_BPT both be defined.
It allows the system to be stopped by pressing the Escape key
(when it would normally add a service credit).  Because of that,
FREE_ONLY is now implied when CONFIG_BPT is enabled.  Escape has
its usual meaning in test mode.

The "break" check occurs once every 16ms, around the same time that
periodic functions are invoked.  As such, you are now inside any
task context when it stops.

Breakpoints manually inserted into the code via the bpt() macro
fall into the shell also.