Support ORKIN and parallel debugging in the same ROM image.
Previously, debugger output would go to either the orkin console
(a fake debugging tool integrated into pinmame) or to the WPC
parallel port (which pinmame also emulates by writing the text into
a file).  The option was controlled at compile-time.

Now, during initialization, if the console is detected, then we
use it; otherwise we fall back to the parallel port.  If debugging
is compiled out completely, then messages are not written at all,
as before.

This change facilitates debugging the same ROMs on Linux and Windows
machines, which have different debugging capabilities.