Don't invoke single_ball_play() at the end of 1-ball play.
This event is now only thrown when multiballs end proper; i.e.
when the balls in play goes from 2 to 1.  This hook allows modes to
stop multiball rules.

Note that if multiple balls drain rapidly, say from 3 to 1, this
still works OK because the logic is invoked on every change in
ball count.  That is, there is a call to device_remove_live() when
it drops to 2, and again when it drops to 1.