Native mode: Improve accuracy of simulation timing.
The coordination between realtime and non-realtime tasks in native mode
is quite different than it is on real hardware, because the host CPU
can run a lot of faster, and it is also not as deterministic.
When measuring signal levels, some anomalies were seen.  This patch
improves some of that.