Support multi-phase solenoid pulses.
If you need a call to sol_request() to vary in its duty cycle throughout the
pulse, follow this procedure:

1. Catch the 'sol_request' event (you only need one of these per machine;
it can handle different solenoids that need such behavior).  At the
point this is called, no other pulse is in progress, and calls can be
made into the driver to manipulate the output line directly.

2. Inspect the value of 'sol_pulsing' to see which solenoid is
being pulsed.  Return TRUE if the value is not one that needs special
processing, and the usual behavior will result.

3. If it matches, call sol_request_async() to start a default pulse,
or sol_req_start_specific() if you need to specify a special duty cycle
or timeout.  This calls return *immediately* before any I/O occurs.

4. Use task_sleep() to wait for some time, before you want to modify
the parameters.

5. Use sol_modify_duty() or sol_modify_timeout() to make changes.
They affect the coil instantly and also return immediately.  Most
importantly, there is no delay between the old pulse and new pulse,
as if you had made two consecutive calls to sol_request().

6. When you are done, call sol_free().  This is important, as it frees
up the driver for other pulses.

7. Last, return FALSE to signal to the caller that you made the pulse
calls yourself.

Note that this functionality is not supported for calls to sol_request_async,
since there is no locking to protect other pulse requests from interrupting.