Support coin door ball save.
This is a feature taken from newer Stern games whereby you can
open the coin door during a game, and it will not go to endball
when the balls drain (due to the flippers losing power).

When the door is closed again, timers will resume and balls will
be added back into play.

The default is OFF for now to match previous behavior; once this
is better tested, this may be changed.