Native mode: Implement better ball tracking.
We now track each individual pinball in a graph of "nodes",
which are positions on the playfield that can queue pinballs.
The coil implementation allows balls to be "kicked" from one
node to the next.  Likewise, actions that previously manipulated
switches directly now operate on the node graph; the switch
results are secondary.  For example, we no longer toggle the
shooter switch on Spacebar, but instead execute a ball move from
the plunger lane to the playfield.