Fix behavior of device_request_empty() for the trough device.
It was marking the trough to not hold onto any balls, which
is right for playfield devices but not for the trough.

Now, empty balls test will attempt to release everything from
the trough, but if the balls fall back in there, they will be
kept and not fire out again.