Pressing Enter during empty balls test will try to empty balls again.
This is useful if the devices couldn't be cleared for some reason
when the menu option was entered.

The display is also fixed to be more meaningful.  It would say
"EMPTY BALLS DONE" only when no balls were detected.  Now, it
would switch to "EMPTY BALLS DONE" as soon as the empty request
comes back.  If you press Enter again, it will switch back to
"EMPTYING BALLS" briefly.