TZ: Don't physically lock 3 balls in the lockup.
The device system will let you do that, but it doesn't currently
know what to do if a new ball needs to be served and the trough is
empty.  In the case of TZ, we never want to hold 3 balls here,
as we can no longer score a lock shot.  Generally, this is the
right approach, but also consider saucers like those on Doctor Who
where a device is allowed to remain full, and it can't be shot anymore.

Now check if the device is full on enter, and if so, don't call
device_lock_ball().  The common code will take care of releasing a
ball from the lock automatically (because count=3 and locks=2).
We could add a lock kickout warning effect here too.