Workaround a make bug with parallel builds.
For some reason the image header is getting generated twice due to
broken dependencies.  While the second imgld instance is running,
some C files are being compiled trying to use it.  This finds
a half-baked version of the file which leads to a compiler error.

As a workaround, added a small delay to try to avoid
parallelization when performing this part of the build.