Simulation: Hide target va_list definition, and use system one.
Previously some simulation files could not include <freewpc.h>
because it included <printf.h>, which brought in a definition for
va_list, etc. that conflicted with the native system definition.
The simulator needs to use the native version when doing logging.
However, the target needs to use its version as usual.

Now, we hide the target version from the simulator code, using the
NATIVE_SYSTEM conditional.  The simulator can include <stdarg.h>.
In one exception case, the pth task library needs to issue
dbprintf(), which is target code, so it needs the target version
and includes it manually.