Add the "deff data" feature.
This is a cleaner way to say which global variables your display
effect needs to reference.  It is a more generic version of the
D_SCORE flag, but for more than just scores.

See include/deffdata.h for the full documentation.  Use deff_data_init()
and deff_data_push() before starting the deff; use deff_data_pull()
inside the deff to get the data that was previously pushed.

Another reason for this is it helps to support future builds on
the Whitestar platform, where global variables can't be referenced
directly, because they are on separate processors.