Fix major bug causing corruption of player 4 data.
The root cause is that the offset to player 4's save region was
not being calculated properly on the 6809, due to 8-bit math
overflow.  As each section size was 64 bytes, the offset to player 4
was 256, but this overflows to 0.

Changed the size of each player's region to a smaller 48 bytes,
which prevents this overflow from happening.

Signed-off-by: Brian Dominy <brian@oddchange.com>