Update TZ-related files from SonnyJim's v1.0-rc branch.
This is not a proper merge, and does not record the version history of
all the changes.  Due to the diverging branch history between v1.0 and
master, this is the easiest way to update the TZ specific files.
It does not include changes that were made to system files.

In theory, the TZ code should be able to work with any system version
(not entirely true, but close).

MergedBy: Brian Dominy <brian@oddchange.com>