Change warm reboot to do a full restart, including start.s code.
Previously it was not clearing memory, which could cause
problems for modules that expected variables to be clean.
The culprit seen was task_init().

This problem exhibited itself by infinite reboot after factory reset,
because the task scheduler wasn't working correctly (it thought
some task slots were in use when they had garbage in them).