Merge filesystem code with checksum code.
A single header is used for both modules, 'filesystem.h', as they
depend on each other now.  csum.c should probably be merged as well.

Added the file type field into the area_csum struct.

The 'csum' field in the area_csum struct is now optional; when NULL
the space in the file table will be used instead.  The old code will
still work for now.

When modules want to switch to the new code, they should:
- Remove the csum declaration.
- Set version = 1.