Rearrange section layout so that callsets are more efficient:
The callset EVENT page is now the same as the MACHINE page.  This means
that CALLSET_ENTRY functions declared in machine code don't require far
calls, which saves about 80 cycles for every function call.  This is a
huge win especially for playfield switches which close very frequently.

It is recommended that machine-specific code that processes switches
go into the MACHINE_PAGE (MACHINE_OBJS) if possible, and use MACHINE2,
MACHINE3, etc. for less commonly used code.  Shot handlers especially
should go into the first MACHINE_PAGE.  It is not an error if this is not
followed, but the system will waste more cycles.

With this change, MACHINE_PAGE has less room than it did before, so existing
ROMs that filled it up may need rearranging.

Also transition code is moved into the same page as that intended for
display effects, since these two are related.