Change the way callset IDs are assigned to optimize them.
IDs are used to debug when the system crashes.  Now, each
event function begins with a multiple of 64, and IDs are
incremented from there.  Previously, IDs just increased from
the top of the file down, regardless of function boundaries...

This lets us optimize by using simple "inc" instructions on
the 6809.  This saves a few cycles, and also a few bytes of
code space.  On TZ, this saved about 1000 bytes!