Move many display effects into separate file from core logic.
The objective here is twofold: first, not all machine have a DMD,
and there is DMD code littered everywhere.  Isolating this will
make porting to alphanumerics easier, or even other future
plans like an LCD...

Second, it helps with Whitestar porting where the DMD is
managed by a separate processor.  There is still tight coupling
at the moment (i.e. the deffs file depends on variables in the
CPU board's address space), but this is a start.

There are more functions to be moved still; these were the easy ones.