Checkin new 'min', or minimal platform.
The intent is to create a bare minimal platform for what is needed
in a pinball game.  Getting this to compile is much easier than
trying to support a "full" platform.  This aids in fixing the
portability problems, and will make it easier to get other real
platforms working in the future.