Import new RTT scheduling tool.
This is not used yet, nor is it ready, but the idea is to replace sched
with something much simpler and to allow for scheduling definitions to
be embedded in the C code, without having separate sched files.  It
works similarly to gencallset in that the C code is scanned for
special directives.