Special lamps can be named in machine config via "%special" syntax:
For example, you can refer to the name of the lamp tagged as
the shoot-again lamp in a lamplist with "%shoot-again".  This
is substituted by genmachine, so the lamplist will only refer to
the real lamp name (LM_SHOOT_AGAIN), not the global name
(MACHINE_SHOOT_AGAIN_LAMP).

With this change, a default shoot again/ball save lamp don't
need to be defined in C code.  Ball save lamp still defaults
to shoot again as before.

Doing this in genmachine makes the lampset emitted for such
lamplists correct, so allocation/free works as it should.