6809: Support object registration for timedmodes, mbmodes, etc.
This is the first of several patches to enable this.  It allows
for additional sections to be linked into the system area.
The list of sections is in section.s.

With this change, code in paged ROM is also allowed to declare
items in any system section (sysrom, timedmodes, mbmodes).
This was not previously possible.  System code is still not allowed
to declare paged functions in the same file.