Assign larger values to MAX_FLAGS, MAX_GLOBAL_FLAGS, MAX_TIMERS.
Since these are no longer restricted to .md files, the values
that genmachine assigns to these are too short.  Proper
definitions require linker variables to see what the actual
size of the sections are, but for now, these definitions should
fix a bunch of problems.  Before, updates to flags & timers were
overwriting adjacent memory with garbage.